Move EventEntry entity to TournamentAggregate
