+++ src/TournamentManagement/TournamentManagement.Domain.UnitTests/TournamentAggregate/EventEntryIdTests.cs
@@ -1,8 +1,9 @@
 ﻿using FluentAssertions;
 using System;
+using TournamentManagement.Domain.TournamentAggregate;
 using Xunit;
 
-namespace TournamentManagement.Domain.UnitTests
+namespace TournamentManagement.Domain.UnitTests.TournamentAggregate
 {
 	public class EventEntryIdTests
 	{
