+++ src/TournamentManagement/TournamentManagement.Domain/TournamentAggregate/EventEntry.cs
@@ -5,9 +5,8 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using TournamentManagement.Domain.Common;
 using TournamentManagement.Domain.PlayerAggregate;
-using TournamentManagement.Domain.TournamentAggregate;
 
-namespace TournamentManagement.Domain
+namespace TournamentManagement.Domain.TournamentAggregate
 {
 	public class EventEntry : Entity<EventEntryId>
 	{
