+++ src/TournamentManagement/TournamentManagement.Domain/TournamentAggregate/EventGenderValidator.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
-using TournamentManagement.Domain.TournamentAggregate;
 
-namespace TournamentManagement.Domain
+namespace TournamentManagement.Domain.TournamentAggregate
 {
 	public static class EventGenderValidator
 	{
